Extract workspace model
